Fixed insert and delete bug in CollectionUsingHashCodesBase --> ready t
- CollectionUsingHashCodesBase big bug in #remove -> but it needs less memory than HashSet and ist also very fast up to 1000-2000 entries (basic idea copied is from jena HashedTripleBunch)
- HashValueMap -> a tiny bit less memory intensive than HashSet and with promising performance -> iterator and stream have to be measured yet (basic implementation follows java HashMap but with less overhead)
- SortedArrayList --> will be renamed to SortedListSet soon --> very promising as very fast set with the memory footprint of a simple ArrayList (which it basically is) (sorting should begin at about 15 entries) --> should be replaced by a hash map implementation somewhere between 350 entries (if one needs fast #add, #remove or #contains but less speed on #iterateor or #stream) and 900/1000 entries (if speed of #iterateor or #stream is more important than #add or #remove)

- open: measure stream.parallel

- Collection#stream.iterator() should always be slower than Collection#iterator (my other thoughts were bullshit)

+++ jena-arq/src/test/java/org/apache/jena/mem/TestGraphMemVariantsBase.java
@@ -100,7 +100,7 @@ public abstract class TestGraphMemVariantsBase {
             }
             stopwatch.stop();
             var memoryAfter = runGcAndGetUsedMemoryInMB();
-            System.out.println("Size of triples in ArrayList: " + (memoryAfter-memoryBefore) + " MB - time to load: " + stopwatch.formatTime());
+            System.out.println("Size of triples in ArrayList: " + (int)(memoryAfter-memoryBefore) + " MB - time to load: " + stopwatch.formatTime());
         }
         return triplesPerGraph;
     }
